+++ framy_generator/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Support for List dependencies of classes annotated with @FramyModel
 * Added `DropdownButton` to Material Component's button page
 * Added Material Components' Toggle page
+* Added Material Components' TextField page
 
 ## [0.1.1]